refactor!: rebrand @pgbo/client → @metadataui/* (closes #52)
The metadata-driven UI contract — URL conventions, response shapes,
status-code semantics — is independent of pgbo. It now lives under a
vendor-neutral namespace so other backends (Spring, Go, hand-rolled
REST) can implement it without taking pgbo as a dependency.

New packages:
- @metadataui/spec@1.0.0 — pure types + URL builders. Zero runtime
  dependencies. Server-and-browser compatible. The contract.
- @metadataui/client@1.0.0 — framework-agnostic HTTP client. Drop-in
  replacement for @pgbo/client (renamed PgboClientError →
  MetadataUiClientError).

Updated packages:
- @pgbo/core@1.1.0 — metadata/types.ts re-exports from @metadataui/spec.
  FieldKind + FilterOption re-exported too. Adds @metadataui/spec dep.
- @pgbo/fastify@3.1.0 — FileResponse re-exports from @metadataui/spec.
  Adds @metadataui/spec dep.

Deprecated:
- @pgbo/client@0.3.0 — re-exports everything from @metadataui/client and
  aliases PgboClientError → MetadataUiClientError. Logs a one-time
  console.warn at module load. Will be removed in the next major.

Migration for apps using @pgbo/client (one find-and-replace):
- import { createClient, PgboClientError } from '@pgbo/client'
+ import { createClient, MetadataUiClientError } from '@metadataui/client'

Apps importing types from @pgbo/core/metadata keep working — those
types are now re-exported from @metadataui/spec but the existing
import paths still resolve.

Versions are pre-applied in package.json so no changeset is included.
First publish of @metadataui/spec + @metadataui/client requires a
manual bootstrap (no Trusted Publisher entry yet); subsequent
releases can run through the workflow.
