+++ packages/metadataui-spec/tsconfig.eslint.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "."
+  },
+  "include": ["src", "tests", "eslint.config.js", "vitest.config.ts"]
+}
